+++ CHANGELOG.md
@@ -15,7 +15,10 @@ When publishing a new release, copy the relevant section on the [Github release 
 - Support for reading CSV files by chunks using pandas.read_csv `chunksize` argument (only for DataStores). 
 This modification is compatible with multi-threading and returns a more accurate progression indicator.
 It is disabled by default. To enable it, use the `allow_chunks` argument or enter `--allow-chunks` in the __Options__ column of the resources worksheet.
-- Support for user-defined file format loading functions with new Excel columns __Read function__ and __Write function__. 
+- Excel builder workbook:
+  - Support for user-defined file format loading functions with new Excel columns __Read function__ and __Write function__.
+  - Field __Limit__ in the _ckan_ sheet to change the number of rows sent per request.
+  - Field __Time between requests__ in the _ckan_ sheet to change the delay between each request (upload/download), in seconds
 - The upload data cleaner can be used to replace empty string values with None for columns indicated as numeric in the field metadata.
 - The admin report can output the storage space used by a dataset in custom fields defined in the data format policy JSON.